scripts/python/requirements.txt
# ZeroMQ
pyzmq